V7.1.0 (#563)
* add: tsconfig-paths-webpack-plugin (#3387)

* fix: set main process.type to "browser" in production (#3370)

According to the documentation, the only valid values for process.type are "browser", "renderer", "worker" and "utility". Currently, running the application with "yarn start" the process.type value in the main process is "browser", but in a packaged app it is "main", so this commit fixes it to be "browser" too.
See https://www.electronjs.org/docs/latest/api/process#processtype-readonly

* chore: 移除tg信息

* fix: 修复baseUrl缺失

* fix: 优化全局任务引起不必要的渲染

* fix: 修复webpack alias引起的编译错误

* chore: 修改发布工作流触发时机

* fix: 修复DetailPage不必要的渲染

* chore: 规范证书

* fix: 修复部分纯ESM模块加载问题

* fix: 修复renderer prod构建

* fix: 修复windows构建符号转义问题

* Bump chalk from 4.1.2 to 5.2.0 (#558)

Bumps [chalk](https://github.com/chalk/chalk) from 4.1.2 to 5.2.0.
- [Release notes](https://github.com/chalk/chalk/releases)
- [Commits](https://github.com/chalk/chalk/compare/v4.1.2...v5.2.0)

---
updated-dependencies:
- dependency-name: chalk
  dependency-type: direct:development
  update-type: version-update:semver-major
...


* Bump query-string from 7.1.3 to 8.1.0 (#556)

Bumps [query-string](https://github.com/sindresorhus/query-string) from 7.1.3 to 8.1.0.
- [Release notes](https://github.com/sindresorhus/query-string/releases)
- [Commits](https://github.com/sindresorhus/query-string/compare/v7.1.3...v8.1.0)

---
updated-dependencies:
- dependency-name: query-string
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* Bump eslint-config-prettier from 8.5.0 to 8.6.0 (#554)

Bumps [eslint-config-prettier](https://github.com/prettier/eslint-config-prettier) from 8.5.0 to 8.6.0.
- [Release notes](https://github.com/prettier/eslint-config-prettier/releases)
- [Changelog](https://github.com/prettier/eslint-config-prettier/blob/main/CHANGELOG.md)
- [Commits](https://github.com/prettier/eslint-config-prettier/compare/v8.5.0...v8.6.0)

---
updated-dependencies:
- dependency-name: eslint-config-prettier
  dependency-type: direct:development
  update-type: version-update:semver-minor
...


* fix: 修改ts-node编译配置

* feat: type module

* fix: 添加tsx

* fix: 移除ts-node

* Bump rimraf from 3.0.2 to 4.1.2 (#560)

Bumps [rimraf](https://github.com/isaacs/rimraf) from 3.0.2 to 4.1.2.
- [Release notes](https://github.com/isaacs/rimraf/releases)
- [Changelog](https://github.com/isaacs/rimraf/blob/main/CHANGELOG.md)
- [Commits](https://github.com/isaacs/rimraf/compare/v3.0.2...v4.1.2)

---
updated-dependencies:
- dependency-name: rimraf
  dependency-type: direct:development
  update-type: version-update:semver-major
...


* Bump ts-jest from 28.0.8 to 29.0.5 (#559)

Bumps [ts-jest](https://github.com/kulshekhar/ts-jest) from 28.0.8 to 29.0.5.
- [Release notes](https://github.com/kulshekhar/ts-jest/releases)
- [Changelog](https://github.com/kulshekhar/ts-jest/blob/main/CHANGELOG.md)
- [Commits](https://github.com/kulshekhar/ts-jest/compare/v28.0.8...v29.0.5)

---
updated-dependencies:
- dependency-name: ts-jest
  dependency-type: direct:development
  update-type: version-update:semver-major
...


* bump: udpate typescript 5.0-beta

* fix: 移除不必要的命令

* fix: 完善.eslintignore

* fix: 回退rimraf

* fix: 移除lint 测试

* feat: 添加fakeUA

* fix: 修复got headers配置问题

* fix: 优化eval间接调用

* fix: 修复近期资讯rowKey

* fix: 修复appCenter遍历key缺失

* feat: 更新移动端UA

* fix: 修复allowpopups失效

* Revert "fix: 优化eval间接调用"

This reverts commit 3fc8c8165e456dfd72bd90748cfb862b8ada2422.

* chore: 移除ts-node、sucrase

* chore: bump eslint-config-erb (#3412)

* chore: bump eslint-config-erb

* chore: drop husky and lint-staged

* chore: bump deps

* chore: bump all deps

* fix: bump again

* fix: bump again

* fix: check delete before delete

* bump: update deps

* bump: 7.1.0

* bump: update electron 23.0.0

* chore: update workflow node to v18

* chore: update copyright

* Bump @types/node from 18.11.9 to 18.11.19 (#565)

Bumps [@types/node](https://github.com/DefinitelyTyped/DefinitelyTyped/tree/HEAD/types/node) from 18.11.9 to 18.11.19.
- [Release notes](https://github.com/DefinitelyTyped/DefinitelyTyped/releases)
- [Commits](https://github.com/DefinitelyTyped/DefinitelyTyped/commits/HEAD/types/node)

---
updated-dependencies:
- dependency-name: "@types/node"
  dependency-type: direct:development
  update-type: version-update:semver-patch
...


---------

+++ .erb/scripts/clean.js
@@ -1,8 +1,9 @@
 import rimraf from 'rimraf';
-import webpackPaths from '../configs/webpack.paths';
+import fs from 'fs';
+import webpackPaths from '../configs/webpack.paths.mjs';
 
 const foldersToRemove = [webpackPaths.distPath, webpackPaths.buildPath, webpackPaths.dllPath];
 
 foldersToRemove.forEach((folder) => {
-  rimraf.sync(folder);
+  if (fs.existsSync(folder)) rimraf.sync(folder);
 });
